version 2: set repeatitions
In this version:
in each number of classes, I set repeatitions. The lowest score in repeatitions will be selected and compared with previous one.
format of input data is optimized, iqtree files and data files are not need to be put together.
the output is the optimal number of classes, all the scores generated in whole algorithm are also listed.